Query Products from the web service and save them in our ContentProvider. Still not displaying them anywhere.

+++ app/src/main/java/com/veskoiliev/asosmini/MainActivityPresenter.java
@@ -8,5 +8,5 @@ public interface MainActivityPresenter {
 
     void onMenuMenClicked();
 
-    void onCategorySelected(int id);
+    void onCategorySelected(long id);
 }
